+++ noodles-bam/src/reader/records.rs
@@ -1,6 +1,6 @@
 use std::io::{self, Read};
 
-use noodles_sam::alignment::Record;
+use noodles_sam::{self as sam, alignment::Record};
 
 use super::Reader;
 
@@ -12,6 +12,7 @@ where
     R: Read,
 {
     reader: &'a mut Reader<R>,
+    header: &'a sam::Header,
     record: Record,
 }
 
@@ -19,9 +20,10 @@ impl<'a, R> Records<'a, R>
 where
     R: Read,
 {
-    pub(super) fn new(reader: &'a mut Reader<R>) -> Records<'_, R> {
+    pub(super) fn new(reader: &'a mut Reader<R>, header: &'a sam::Header) -> Self {
         Self {
             reader,
+            header,
             record: Record::default(),
         }
     }
@@ -34,7 +36,7 @@ where
     type Item = io::Result<Record>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        match self.reader.read_record(&mut self.record) {
+        match self.reader.read_record(self.header, &mut self.record) {
             Ok(0) => None,
             Ok(_) => Some(Ok(self.record.clone())),
             Err(e) => Some(Err(e)),