+++ noodles-bam/src/reader/unmapped_records.rs
@@ -1,7 +1,7 @@
 use std::io::{self, Read};
 
 use noodles_bgzf as bgzf;
-use noodles_sam::alignment::Record;
+use noodles_sam::{self as sam, alignment::Record};
 
 use super::Reader;
 
@@ -13,6 +13,7 @@ where
     R: Read,
 {
     reader: &'a mut Reader<bgzf::Reader<R>>,
+    header: sam::Header,
     record: Record,
 }
 
@@ -23,6 +24,7 @@ where
     pub(crate) fn new(reader: &'a mut Reader<bgzf::Reader<R>>) -> Self {
         Self {
             reader,
+            header: sam::Header::default(),
             record: Record::default(),
         }
     }
@@ -36,7 +38,7 @@ where
 
     fn next(&mut self) -> Option<Self::Item> {
         loop {
-            match self.reader.read_record(&mut self.record) {
+            match self.reader.read_record(&self.header, &mut self.record) {
                 Ok(0) => return None,
                 Ok(_) => {
                     if self.record.flags().is_unmapped() {