+++ noodles-bam/CHANGELOG.md
@@ -8,7 +8,8 @@
     rather than a reference sequence dictionary
     (`sam::header::ReferenceSequences`).
 
-  * bam/reader: Change `Reader::query` to require a header (`sam::Header`).
+  * bam/reader: Require `Reader::read_record` and `Reader::records` to receive
+    a header (`sam::Header`).
 
 ## 0.26.0 - 2023-02-03